pyproject.toml
[project]
name = "ztx-calendar-bot"
version = "1.2.0"
description = "Calendar Telegram Bot"
readme = "README.md"
authors = [
    { name = "t.me/zxcdiana" }
]
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.23.0",
    "aiogram-i18n>=1.4",
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "cryptg>=0.5.2",
    "fluent-runtime>=0.4.0",
    "meval",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-utc>=0.14.0",
    "telethon",
    "typer-slim>=0.21.1",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "ftl-extract>=0.9.0",
]

[tool.uv.sources]
meval = { url = "https://github.com/zxcdiana/meval/archive/master.zip" }
telethon = { url = "https://github.com/LonamiWebs/telethon/archive/v1.zip" }

[tool.ruff]
exclude = [
    "migrations/",
    "app/_stub.pyi"
]

[tool.pyright]
typeCheckingMode = "basic"
reportOptionalMemberAccess = "none"
